+++ src/Sentry.AspNetCore/Extensions/DependencyInjection/ServiceCollectionExtensions.cs
@@ -4,6 +4,10 @@ using Sentry.AspNetCore;
 using Sentry.Extensibility;
 using Sentry.Extensions.Logging.Extensions.DependencyInjection;
 
+#if !NETSTANDARD
+using Microsoft.Extensions.Http;
+#endif
+
 // ReSharper disable once CheckNamespace -- Discoverability
 namespace Microsoft.Extensions.DependencyInjection
 {
@@ -20,15 +24,21 @@ namespace Microsoft.Extensions.DependencyInjection
         /// <returns></returns>
         public static ISentryBuilder AddSentry(this IServiceCollection services)
         {
-            _ = services.AddSingleton<ISentryEventProcessor, AspNetCoreEventProcessor>();
+            services.AddSingleton<ISentryEventProcessor, AspNetCoreEventProcessor>();
             services.TryAddSingleton<IUserFactory, DefaultUserFactory>();
 
-            _ = services
+            services
                     .AddSingleton<IRequestPayloadExtractor, FormRequestPayloadExtractor>()
                     // Last
                     .AddSingleton<IRequestPayloadExtractor, DefaultRequestPayloadExtractor>();
 
-            _ = services.AddSentry<SentryAspNetCoreOptions>();
+            services.AddSentry<SentryAspNetCoreOptions>();
+
+#if !NETSTANDARD2_0
+            // Custom handler for HttpClientFactory.
+            // Must be singleton: https://github.com/getsentry/sentry-dotnet/issues/785
+            services.AddSingleton<IHttpMessageHandlerBuilderFilter, SentryHttpMessageHandlerBuilderFilter>();
+#endif
 
             return new SentryAspNetCoreBuilder(services);
         }