+++ test/Sentry.AspNetCore.Tests/SentryTracingMiddlewareTests.cs
@@ -1,5 +1,6 @@
 ﻿#if !NETCOREAPP2_1
 using System;
+using System.Net.Http;
 using System.Threading.Tasks;
 using FluentAssertions;
 using Microsoft.AspNetCore.Builder;
@@ -29,6 +30,7 @@ namespace Sentry.AspNetCore.Tests
             });
 
             var server = new TestServer(new WebHostBuilder()
+                .UseDefaultServiceProvider(di => di.ValidateScopes = true)
                 .UseSentry()
                 .ConfigureServices(services =>
                 {
@@ -79,6 +81,7 @@ namespace Sentry.AspNetCore.Tests
             });
 
             var server = new TestServer(new WebHostBuilder()
+                .UseDefaultServiceProvider(di => di.ValidateScopes = true)
                 .UseSentry()
                 .ConfigureServices(services =>
                 {
@@ -96,7 +99,7 @@ namespace Sentry.AspNetCore.Tests
                     {
                         routes.Map("/person/{id}", _ =>
                         {
-                            transaction = hub.ScopeManager.GetCurrent().Key.Transaction;
+                            transaction = hub.GetSpan() as ITransaction;
                             return Task.CompletedTask;
                         });
                     });
@@ -111,6 +114,63 @@ namespace Sentry.AspNetCore.Tests
             // Assert
             transaction.Should().NotBeNull();
             transaction?.Name.Should().Be("GET /person/{id}");
+        }
+
+        [Fact]
+        public async Task Transaction_is_started_automatically_from_incoming_trace_header()
+        {
+            // Arrange
+            ITransaction transaction = null;
+
+            var sentryClient = Substitute.For<ISentryClient>();
+
+            var hub = new Internal.Hub(sentryClient, new SentryOptions
+            {
+                Dsn = DsnSamples.ValidDsnWithoutSecret
+            });
+
+            var server = new TestServer(new WebHostBuilder()
+                .UseDefaultServiceProvider(di => di.ValidateScopes = true)
+                .UseSentry()
+                .ConfigureServices(services =>
+                {
+                    services.AddRouting();
+
+                    services.RemoveAll(typeof(Func<IHub>));
+                    services.AddSingleton<Func<IHub>>(() => hub);
+                })
+                .Configure(app =>
+                {
+                    app.UseRouting();
+                    app.UseSentryTracing();
+
+                    app.UseEndpoints(routes =>
+                    {
+                        routes.Map("/person/{id}", _ =>
+                        {
+                            transaction = hub.GetSpan() as ITransaction;
+                            return Task.CompletedTask;
+                        });
+                    });
+                })
+            );
+
+            var client = server.CreateClient();
+
+            // Act
+            using var request = new HttpRequestMessage(HttpMethod.Get, "/person/13")
+            {
+                Headers = {{"sentry-trace", "75302ac48a024bde9a3b3734a82e36c8-1000000000000000-0"}}
+            };
+
+            await client.SendAsync(request);
+
+            // Assert
+            transaction.Should().NotBeNull();
+            transaction?.Name.Should().Be("GET /person/{id}");
+            transaction.TraceId.Should().Be(SentryId.Parse("75302ac48a024bde9a3b3734a82e36c8"));
+            transaction.ParentSpanId.Should().Be(SpanId.Parse("1000000000000000"));
+            transaction.IsSampled.Should().BeFalse();
         }
     }
 }